% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish data.}
\format{A tibble of 250 rows and 6 columns.  Good for demonstrating zero-inflated models.
\describe{
  \item{nofish}{I've never seen this explained.  Originally 0 and 1, 0 is
  equivalent to \code{livebait == 'yes'}, so I think it is whether the
  primary motivation of the camping trip is for fishing or not.  So
  \code{no?fish == 'yes'} means the trip was not just for fishing.}
  \item{livebait}{whether live bait was used or not}
  \item{camper}{whether or not they brought a camper}
  \item{persons}{how many total persons on the trip}
  \item{child}{how many children present}
  \item{count}{number of fish count}
}}
\source{
\url{https://stats.idre.ucla.edu/r/dae/zip/}
}
\usage{
fish
}
\description{
State wildlife biologists want to model how many fish are being caught by
fishermen at a state park.
}
\details{
Visitors are asked how long they stayed, how many people were in
  the group, were there children in the group and how many fish were caught.
  Some visitors do not fish, but there is no data on whether a person fished
  or not. Some visitors who did fish did not catch any fish so there are
  excess zeros in the data because of the people that did not fish.
}
\examples{
library(noiris)
str(fish)
}
\keyword{datasets}
