% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder_2019}
\alias{gapminder_2019}
\title{Gapminder data.}
\format{The main data frame \code{gapminder} has 40,953 rows and 7 variables:
\describe{
  \item{country}{factor with 142 levels}
  \item{continent}{factor with 5 levels}
  \item{year}{ranges from 1800 to 2018}
  \item{lifeExp}{life expectancy at birth, in years}
  \item{pop}{population}
  \item{gdpPercap}{GDP per capita (US$, inflation-adjusted)}
  \item{giniPercap}{Gini coefficient of inequality}
  }}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
gapminder_2019
}
\description{
A recent download of the Gapminder data on life expectancy, GDP per capita, Gini coefficient, and
population by country.
}
\details{
While the data goes back to 1800, most of it prior to a couple of
  decades ago is either filler or probably nonsense.  See the
  \link[gapminder]{gapminder} package for a smaller version.
}
\examples{
library(noiris)
str(gapminder_2019)
gapminder_2019
summary(gapminder_2019)

if (require("dplyr")) {
gapminder_2019 \%>\%
  filter(year == 2007) \%>\%
  group_by(continent) \%>\%
  summarise(lifeExp = median(lifeExp))
}

}
\seealso{
\link[gapminder]{gapminder}
}
\keyword{datasets}
