% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_apps.R
\docType{data}
\name{google_apps}
\alias{google_apps}
\title{Google apps data}
\format{The nested main data frame \code{google_apps} has 9707 rows and 14
  variables:

\describe{
  \item{app}{Application name}
  \item{category}{Category the app belongs to}
  \item{rating}{Overall user rating of the app}
  \item{reviews}{Number of user reviews for the app}
  \item{size_in_MB}{Size of the app in megabytes}
  \item{installs}{Number of user downloads/installs for the app}
  \item{type}{Paid or Free}
  \item{price}{Price of the app}
  \item{content_rating}{Age group the app is targeted at - Children / Mature
  21+ / Adult}
  \item{genres}{An app can belong to multiple genres (apart from its main
  category). For example, a musical family game will belong to Music, Game,
  Family genres.}
  \item{last_updated}{Date when the app was last updated on Play Store}
  \item{current_ver}{Current version of the app available on Play Store}
  \item{android_ver}{Min required Android version}
  \item{greviews}{A list column. See details.}
  }}
\source{
Data set created by \href{https://www.kaggle.com/lava18/google-play-store-apps/}{Lavanya Gupta}
}
\usage{
google_apps
}
\description{
Ratings and other information for Google Play Store apps.
}
\details{
This data set is joins two data from the Kaggle link.  The data was
  scraped and to some extent cleaned before presented to Kaggle.  This data
  has alterations to make the \code{last_updated} an actual date column
  (range 2010-2018), and the size a numeric column after stripping the M
  (megabyte) and k (kilobyte). The final column is a list column of actual
  reviews that can be used for text analysis, sentiment analysis and similar.
  Each data set within \code{greviews} contains the following columns:

\describe{
  \item{Translated_Review}{User review (preprocessed and translated to
  English)}
  \item{Sentiment}{Positive/Negative/Neutral (preprocessed)}
  \item{Sentiment_Polarity}{Sentiment polarity score}
  \item{Sentiment_Subjectivity}{Sentiment subjectivity score}
  }
}
\note{
License is \href{https://www.kaggle.com/lava18/google-play-store-apps#license.txt}{CC-BY}
}
\examples{
library(noiris)
str(google_apps, 1)



}
\keyword{datasets}
