% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movielens.R
\docType{data}
\name{movielens}
\alias{movielens}
\title{MovieLens data.}
\format{A data frame of 1 million rows and 27 columns regarding 184164 users
  and 21745 distinct movies.
\describe{
  \item{user_id}{Numeric user id}
  \item{movie_id}{Numeric movie id}
  \item{rating}{Numeric favorability rating from .5 to 5 by .5.}
  \item{timestamp}{Time the rating was given}
  \item{title}{A clean version of the title. See details.}
  \item{year}{The year of the title}
  \item{genres}{A list column whose elements are character vectors of genres applicable to the title}
  \item{action}{The genres as a binary indicator}
  \item{adventure}{The genres as a binary indicator}
  \item{drama}{The genres as a binary indicator}
  \item{scifi}{The genres as a binary indicator}
  \item{thriller}{The genres as a binary indicator}
  \item{crime}{The genres as a binary indicator}
  \item{romance}{The genres as a binary indicator}
  \item{animation}{The genres as a binary indicator}
  \item{children}{The genres as a binary indicator}
  \item{comedy}{The genres as a binary indicator}
  \item{fantasy}{The genres as a binary indicator}
  \item{imax}{The genres as a binary indicator}
  \item{horror}{The genres as a binary indicator}
  \item{western}{The genres as a binary indicator}
  \item{war}{The genres as a binary indicator}
  \item{mystery}{The genres as a binary indicator}
  \item{musical}{The genres as a binary indicator}
  \item{film_noir}{The genres as a binary indicator}
  \item{documentary}{The genres as a binary indicator}
  \item{no_genre}{The genres as a binary indicator}
  }}
\source{
\url{https://grouplens.org/datasets/movielens/}
}
\usage{
movielens
}
\description{
A sample of the recent MovieLens data
}
\details{
This is a random sample of 1 million observations from the MovieLens data (not the 1 million data set version listed at the ),
  often used as a machine learning benchmark. Differences from the raw data include separation
  of title and year, and adding the genres as indicator variables.  See the
  \href{http://files.grouplens.org/datasets/movielens/ml-20m-README.html}{associated
  ReadMe file} for more details.
}
\examples{
library(noiris)
str(movielens)

}
\keyword{datasets}
