% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiris-package.R
\docType{package}
\name{noiris}
\alias{noiris}
\alias{noiris-package}
\title{noiris: Data sets}
\description{
This package is primarily to provide data that is more relevant to
 actual data people would typically come across in the wild, or is simply
 more interesting (at least to me).  Far too often examples use iris, mtcars,
 etc. for convenience, but actually are inconvenient for demonstrating real
 data and modeling problems, or are too small to even be realistic.  This
 package will provide larger and at some point messier data.
 @details Right now it has:

\describe{
  \item{\code{gapminder}}{a 2019 pull from \url{http://www.gapminder.org/data/}}
  \item{\code{starwars}}{a variety of cleaned data sets produced using the \link[rwars]{rwars} package.}
  }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-clark/fckiris}
  \item Report bugs at \url{https://github.com/m-clark/fckiris/issues}
}

}
\author{
\strong{Maintainer}: Michael Clark \email{micl@umich.edu}

}
\keyword{internal}
