% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{S&P 500 data.}
\format{The nested main data frame \code{sp500} has 505 rows and 10
  variables:
\describe{
  \item{ticker}{}
  \item{company}{}
  \item{sector}{Global Industry Classification Standard}
  \item{sub_industry}{Global Industry Classification Standard}
  \item{address}{City, State}
  \item{date_added}{}
  \item{CIK}{A Central Index Key or CIK number is a number given to an individual, company, or foreign government by the United States Securities and Exchange Commission. The number is used to identify its filings in several online databases, including EDGAR.}
  \item{year_founded}{Year company was founded.}
  \item{pricing_data}{A list column. See details.}
  }}
\usage{
sp500
}
\description{
Includ
}
\details{

}
\examples{
library(noiris)
str(sp500, 1)



}
\keyword{datasets}
