% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_risk.R
\docType{data}
\name{water_risk}
\alias{water_risk}
\alias{water_risk_province}
\title{Water risk data}
\format{Two data frames. One is at the country level (n = 189), while the
  other is at the province level (n=3311).

\describe{
  \item{gid_1}{ISO A3 country code based on GADM.}
  \item{name_0}{National or political entity name based on GADM.}
  \item{name_1}{Sub-national or political entity name based on GADM.}
  \item{iso_a3}{ISO country code.}
  \item{iso_n3}{ISO numeric code.}
  \item{score}{water risk score mapped to a [0-5] scale.}
  \item{score_ranked}{ranking of score.}
  \item{label}{5 category label of score, including 'NoData'.}
  \item{un_region}{UN regions: Africa, Americas, Asia, Europe, Oceania.}
  \item{wb_region}{World Bank regions.}
  \item{population_2019_million}{Estimated country population 2019 in millions.}
  }}
\source{
\href{https://github.com/rutgerhofste/aqueduct30_country_rankings_data_download/blob/master/metadata.md}{Data link}.
\href{https://www.wri.org/resources/data-sets/aqueduct-30-country-rankings}{Main website}.
\href{https://www.nytimes.com/interactive/2019/08/06/climate/world-water-stress.html}{NY Times article}.
}
\usage{
water_risk
}
\description{
From the website: Water is essential to the progress of human societies. It
is required for a healthy environment and a thriving economy. Food
production, electricity generation, and manufacturing, among other things,
all depend on it. However, many decision-makers lack the technical expertise
to fully understand hydrological information. In response to growing concerns
from the private sector and other actors about water availability, water
quality, climate change, and increasing demand, WRI applied the composite
index approach as a robust communication tool to translate hydrological data
into intuitive indicators of water-related risks.
}
\details{
More details on the variables can be found at the links below. The
  weight variables and related have been removed, but the raw data may be
  found at the data link below, and the data which was specifically used for
  this can be found
  \href{https://github.com/m-clark/noiris/blob/master/data-raw/water-risk/aqueduct-30-country-rankings.xlsx}{here}.
}
\note{
License is MIT
}
\examples{
library(noiris)
str(water_risk)
str(water_risk_province)
}
\references{
Hofste, R., S. Kuzma, S. Walker, E.H. Sutanudjaja, et. al. 2019.
  “Aqueduct 3.0: Updated Decision- Relevant Global Water Risk Indicators.”
  Technical Note. Washington, DC: World Resources Institute. Available online
  at: https://www.wri.org/publication/aqueduct-30.
}
\keyword{datasets}
