% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish data.}
\format{The main data frame \code{gapminder} has 40,953 rows and 7 variables:
\describe{
  \item{nofish}{?}
  \item{livebait}{whether live bait was used or not}
  \item{camper}{ranges from 1800 to 2018}
  \item{persons}{how many total persons on the trip}
  \item{child}{how many children present}
  \item{count}{number of fish count}
}}
\source{
\url{https://stats.idre.ucla.edu/r/dae/zip/}
}
\usage{
fish
}
\description{
State wildlife biologists want to model how many fish are being caught by
fishermen at a state park.
}
\details{
Visitors are asked how long they stayed, how many people were in
  the group, were there children in the group and how many fish were caught.
  Some visitors do not fish, but there is no data on whether a person fished
  or not. Some visitors who did fish did not catch any fish so there are
  excess zeros in the data because of the people that did not fish.
}
\examples{
library(noiris)
str(fish)
}
\keyword{datasets}
