% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiris-package.R
\docType{package}
\name{noiris}
\alias{noiris}
\alias{noiris-package}
\title{noiris: Data sets}
\description{
The goal of this package is primarily to provide data that is
  more relevant to the kind people would more typically come across in the wild,
  or is simply more interesting (at least to me).  Far too often examples use
  iris, mtcars, etc. for convenience, but these actually are inconvenient for
  demonstrating real data and modeling problems, or are too small to be very
  realistic examples of everyday data.  This package will provide larger and
  at some point messier data, that is better named, better documented, and
  would be useful across a variety of modeling contexts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-clark/fckiris}
  \item Report bugs at \url{https://github.com/m-clark/fckiris/issues}
}

}
\author{
\strong{Maintainer}: Michael Clark \email{micl@umich.edu}

}
\keyword{internal}
